Moves enumeration_reflection to separate file
