+++ lib/enumerations/reflection.rb
@@ -1,0 +1,11 @@
+module Enumeration
+  EnumerationReflection = Struct.new(:name, :options) do
+    def class_name
+      options[:class_name]
+    end
+
+    def foreign_key
+      options[:foreign_key]
+    end
+  end
+end
